requirements.txt
asgiref==3.2.10
Django==3.0.8
djangorestframework==3.11.0
pytz==2020.1
selenium==3.141.0
sqlparse==0.3.1
urllib3==1.25.9
